# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRISM Engine API",
    "description": "Multi-dimensional STEAM career guidance: psychometric + financial vectorisation, financial constraint solver, parent-student conflict index, market blending, SWOT and roadmaps.\n\nAll /api/v1 responses use the envelope {success, data, error, meta}. MOCK_MODE is currently **True**.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuthResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuthResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TokenPair_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UserOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assessments/{instrument_code}/questions": {
      "get": {
        "tags": [
          "assessment"
        ],
        "summary": "Questions",
        "operationId": "questions_api_v1_assessments__instrument_code__questions_get",
        "parameters": [
          {
            "name": "instrument_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Question__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments/{instrument_code}/submit": {
      "post": {
        "tags": [
          "assessment"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_assessments__instrument_code__submit_post",
        "parameters": [
          {
            "name": "instrument_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SubmitResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "compat (team aliases)"
        ],
        "summary": "Login",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Answer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Likert: '1'..'5'; MCQ: option key"
          },
          "response_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Ms",
            "description": "Optional; very fast answers are down-weighted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_id",
          "value"
        ],
        "title": "Answer"
      },
      "AuthResult": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPair"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "tokens"
        ],
        "title": "AuthResult"
      },
...
      "CompatLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "CompatLoginRequest"
      },
      "CompatLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user_id",
          "role"
        ],
        "title": "CompatLoginResponse"
      },
...
      "Envelope_AuthResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[AuthResult]"
      },
...
      "Envelope_SubmitResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmitResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[SubmitResult]"
      },
...
      "Envelope_TokenPair_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenPair"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[TokenPair]"
      },
...
      "Envelope_UserOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[UserOut]"
      },
...
      "Envelope_list_Question__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Question"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[list[Question]]"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "v1/0.1.0"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "description": "true when the payload is a MOCK_MODE fixture",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "Meta"
      },
...
      "Question": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "instrument_code": {
            "type": "string",
            "title": "Instrument Code"
          },
          "dimension": {
            "type": "string",
            "title": "Dimension",
            "description": "One of the canonical dimensions (see /system/methodology)"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/QuestionOption"
            },
            "type": "array",
            "title": "Options"
          },
          "reverse_scored": {
            "type": "boolean",
            "title": "Reverse Scored",
            "default": false
          },
          "order": {
            "type": "integer",
            "title": "Order"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "instrument_code",
          "dimension",
          "type",
          "prompt",
          "options",
          "order"
        ],
        "title": "Question"
      },
      "QuestionOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "QuestionOption"
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "likert5",
          "mcq"
        ],
        "title": "QuestionType"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "full_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "admin cannot self-register"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth",
            "description": "Required for students (minor consent check)"
          },
          "preferred_language": {
            "type": "string",
            "maxLength": 8,
            "title": "Preferred Language",
            "default": "en"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name",
          "role"
        ],
        "title": "RegisterRequest"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "student",
          "parent",
          "educator",
          "admin"
        ],
        "title": "Role"
      },
...
      "SubmitAnswersRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "maxItems": 300,
            "minItems": 1,
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "SubmitAnswersRequest"
      },
      "SubmitResult": {
        "properties": {
          "instrument_code": {
            "type": "string",
            "title": "Instrument Code"
          },
          "submission_id": {
            "type": "string",
            "title": "Submission Id"
          },
          "scored": {
            "items": {
              "$ref": "#/components/schemas/TraitScore"
            },
            "type": "array",
            "title": "Scored"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags",
            "examples": [
              [
                "straight_lining_detected"
              ]
            ]
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument_code",
          "submission_id",
          "scored",
          "submitted_at"
        ],
        "title": "SubmitResult"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access-token lifetime in seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair"
      },
...
      "TraitScore": {
        "properties": {
          "dimension": {
            "type": "string",
            "title": "Dimension"
          },
          "raw": {
            "type": "number",
            "title": "Raw"
          },
          "normalized": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Normalized",
            "description": "Normalised to [0, 1]"
          },
          "percentile": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentile"
          },
          "reliability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Reliability",
            "description": "Item consistency / coverage for this dimension"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "imputed": {
            "type": "boolean",
            "title": "Imputed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dimension",
          "raw",
          "normalized",
          "reliability",
          "answered"
        ],
        "title": "TraitScore"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_minor": {
            "type": "boolean",
            "title": "Is Minor"
          },
          "consent_status": {
            "type": "string",
            "title": "Consent Status",
            "description": "not_required | pending | granted | revoked"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_minor",
          "consent_status",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}